core/requirements.txt
# MCP Server Dependencies (minimal subset)
# Install: pip install -r requirements.txt

mcp>=1.0.0
duckdb>=1.2.0
requests>=2.31.0
python-dotenv>=1.0.1
mygeotab>=0.9.0
google-generativeai>=0.8.0
httpx>=0.27.0
